status_notification prep - brain.Ponderation -> models.Ponderation - ntfy.py -> notification.py - prep notification.status_notification function
